requirements.txt
psycopg2-binary
python-dotenv
google-generativeai
streamlit
pypdf
